Simplify frame drawing (breaking change)
The previous API kind of exposed the inner workings of the library. It
was also more annoying to use, and I want an easy API. If it turns out
the old API was really needed for something, I can revert this I guess.
The new API may allow for more optimization in the future.